[UITests]Add Xamarin.Forms.Core.UITests.Shared
