+++ Xamarin.Forms.Core.UITests.Shared/Queries.cs
