+++ Xamarin.Forms.Core.UITests.Shared/Remotes/RemoteFactory.cs
