+++ Xamarin.Forms.Core.UITests.Shared/Tests/ActivityIndicatorUITests.cs
