+++ Xamarin.Forms.Core.UITests.Shared/Tests/AppearingUITests.cs
