+++ Xamarin.Forms.Core.UITests.Shared/Tests/BoxViewUITests.cs
