+++ Xamarin.Forms.Core.UITests.Shared/Tests/DatePickerUITests.cs
