+++ Xamarin.Forms.Core.UITests.Shared/Tests/EditorUITests.cs
