+++ Xamarin.Forms.Core.UITests.Shared/Tests/EntryUITests.cs
