+++ Xamarin.Forms.Core.UITests.Shared/Tests/FrameUITests.cs
