+++ Xamarin.Forms.Core.UITests.Shared/Tests/ImageUITests.cs
