+++ Xamarin.Forms.Core.UITests.Shared/Tests/LabelUITests.cs
