+++ Xamarin.Forms.Core.UITests.Shared/Tests/Legacy-CellsUITests.cs
