+++ Xamarin.Forms.Core.UITests.Shared/Tests/PickerUITests.cs
