+++ Xamarin.Forms.Core.UITests.Shared/Tests/ProgressBarUITests.cs
