+++ Xamarin.Forms.Core.UITests.Shared/Tests/RootGalleryUITests.cs
