+++ Xamarin.Forms.Core.UITests.Shared/Tests/SearchBarUITests.cs
