+++ Xamarin.Forms.Core.UITests.Shared/Tests/SliderUITests.cs
