+++ Xamarin.Forms.Core.UITests.Shared/Tests/StepperUITests.cs
