+++ Xamarin.Forms.Core.UITests.Shared/Tests/SwitchUITests.cs
