+++ Xamarin.Forms.Core.UITests.Shared/Tests/TimePickerUITests.cs
