+++ Xamarin.Forms.Core.UITests.Shared/Tests/ViewUITests.cs
