+++ Xamarin.Forms.Core.UITests.Shared/UITestCategories.cs
