+++ Xamarin.Forms.Core.UITests.Shared/Utilities/AppExtensions.cs
