+++ Xamarin.Forms.Core.UITests.Shared/Utilities/Drag.cs
