+++ Xamarin.Forms.Core.UITests.Shared/Utilities/Gestures.cs
