+++ Xamarin.Forms.Core.UITests.Shared/Utilities/Logger.cs
