+++ Xamarin.Forms.Core.UITests.Shared/Utilities/NumericExtensions.cs
