+++ Xamarin.Forms.Core.UITests.Shared/Utilities/ParsingUtils.cs
