+++ Xamarin.Forms.Core.UITests.Shared/Utilities/UITestCustomExceptions.cs
