+++ Xamarin.Forms.Core.UITests.Shared/Xamarin.Forms.Core.UITests.projitems
@@ -1,0 +1,59 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
+    <HasSharedItems>true</HasSharedItems>
+    <SharedGUID>e175485b-3c8c-47d7-8dd5-f7fed627eb25</SharedGUID>
+  </PropertyGroup>
+  <PropertyGroup Label="Configuration">
+    <Import_RootNamespace>Xamarin.Forms.Core.UITests.Shared</Import_RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)BaseTestFixture.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)PlatformQueries.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Queries.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Remotes\BaseViewContainerRemote.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Remotes\EventViewContainerRemote.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Remotes\LayeredViewContainerRemote.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Remotes\RemoteFactory.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Remotes\StateViewContainerRemote.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Remotes\ViewContainerRemote.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\ActionSheetUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\ActivityIndicatorUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\AppearingUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\AutomationIDUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\BoxViewUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\ButtonUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\ContextActionsUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\DatePickerUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\DisplayAlertUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\EditorUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\EntryUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\FrameUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\ImageUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\LabelUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\Legacy-CellsUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\Legacy-UnevenListTests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\PickerUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\ProgressBarUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\RootGalleryUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\ScrollViewUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\SearchBarUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\SliderUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\StepperUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\SwitchUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\TimePickerUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\ToolbarItemTests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\ViewUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tests\WebViewUITests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UITestCategories.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utilities\AppExtensions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utilities\Drag.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utilities\Gestures.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utilities\Logger.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utilities\NumericExtensions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utilities\ParsingUtils.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utilities\UITestCustomExceptions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utilities\ViewInspector.cs" />
+  </ItemGroup>
+</Project>
