+++ Xamarin.Forms.Core.Windows.UITests/Xamarin.Forms.Core.Windows.UITests.csproj
@@ -52,140 +52,6 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\BaseTestFixture.cs">
-      <Link>BaseTestFixture.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Queries.cs">
-      <Link>Queries.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Remotes\BaseViewContainerRemote.cs">
-      <Link>Remotes\BaseViewContainerRemote.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Remotes\EventViewContainerRemote.cs">
-      <Link>Remotes\EventViewContainerRemote.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Remotes\LayeredViewContainerRemote.cs">
-      <Link>Remotes\LayeredViewContainerRemote.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Remotes\RemoteFactory.cs">
-      <Link>Remotes\RemoteFactory.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Remotes\StateViewContainerRemote.cs">
-      <Link>Remotes\StateViewContainerRemote.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Remotes\ViewContainerRemote.cs">
-      <Link>Remotes\ViewContainerRemote.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\ActionSheetUITests.cs">
-      <Link>Tests\ActionSheetUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\ActivityIndicatorUITests.cs">
-      <Link>Tests\ActivityIndicatorUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\AutomationIDUITests.cs">
-      <Link>Tests\AutomationIDUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\BoxViewUITests.cs">
-      <Link>Tests\BoxViewUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\ButtonUITests.cs">
-      <Link>Tests\ButtonUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\ContextActionsUITests.cs">
-      <Link>Tests\ContextActionsUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\DatePickerUITests.cs">
-      <Link>Tests\DatePickerUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\DisplayAlertUITests.cs">
-      <Link>Tests\DisplayAlertUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\EditorUITests.cs">
-      <Link>Tests\EditorUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\EntryUITests.cs">
-      <Link>Tests\EntryUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\FrameUITests.cs">
-      <Link>Tests\FrameUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\ImageUITests.cs">
-      <Link>Tests\ImageUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\LabelUITests.cs">
-      <Link>Tests\LabelUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\Legacy-CellsUITests.cs">
-      <Link>Tests\Legacy-CellsUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\Legacy-UnevenListTests.cs">
-      <Link>Tests\Legacy-UnevenListTests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\PickerUITests.cs">
-      <Link>Tests\PickerUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\ProgressBarUITests.cs">
-      <Link>Tests\ProgressBarUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\RootGalleryUITests.cs">
-      <Link>Tests\RootGalleryUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\ScrollViewUITests.cs">
-      <Link>Tests\ScrollViewUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\SearchBarUITests.cs">
-      <Link>Tests\SearchBarUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\SliderUITests.cs">
-      <Link>Tests\SliderUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\StepperUITests.cs">
-      <Link>Tests\StepperUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\SwitchUITests.cs">
-      <Link>Tests\SwitchUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\TimePickerUITests.cs">
-      <Link>Tests\TimePickerUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\ToolbarItemTests.cs">
-      <Link>Tests\ToolbarItemTests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\ViewUITests.cs">
-      <Link>Tests\ViewUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Tests\WebViewUITests.cs">
-      <Link>Tests\WebViewUITests.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\UITestCategories.cs">
-      <Link>UITestCategories.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Utilities\AppExtensions.cs">
-      <Link>Utilities\AppExtensions.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Utilities\Drag.cs">
-      <Link>Utilities\Drag.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Utilities\Gestures.cs">
-      <Link>Utilities\Gestures.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Utilities\Logger.cs">
-      <Link>Utilities\Logger.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Utilities\NumericExtensions.cs">
-      <Link>Utilities\NumericExtensions.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Utilities\ParsingUtils.cs">
-      <Link>Utilities\ParsingUtils.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Utilities\UITestCustomExceptions.cs">
-      <Link>Utilities\UITestCustomExceptions.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Forms.Core.iOS.UITests\Utilities\ViewInspector.cs">
-      <Link>Utilities\ViewInspector.cs</Link>
-    </Compile>
-    <Compile Include="App.cs" />
-    <Compile Include="PlatformQueries.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -213,6 +79,7 @@
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
   </ItemGroup>
   <Import Project="..\Xamarin.Forms.Controls.Issues\Xamarin.Forms.Controls.Issues.Shared\Xamarin.Forms.Controls.Issues.Shared.projitems" Label="Shared" />
+  <Import Project="..\Xamarin.Forms.Core.UITests.Shared\Xamarin.Forms.Core.UITests.projitems" Label="Shared" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
