+++ Xamarin.Forms.Core.iOS.UITests/Xamarin.Forms.Core.iOS.UITests.csproj
@@ -64,53 +64,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BaseTestFixture.cs" />
-    <Compile Include="UITestCategories.cs" />
-    <Compile Include="PlatformQueries.cs" />
-    <Compile Include="Remotes\BaseViewContainerRemote.cs" />
-    <Compile Include="Remotes\EventViewContainerRemote.cs" />
-    <Compile Include="Remotes\LayeredViewContainerRemote.cs" />
-    <Compile Include="Remotes\ViewContainerRemote.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Queries.cs" />
-    <Compile Include="Remotes\StateViewContainerRemote.cs" />
-    <Compile Include="Remotes\RemoteFactory.cs" />
-    <Compile Include="Tests\ContextActionsUITests.cs" />
-    <Compile Include="Tests\ActivityIndicatorUITests.cs" />
-    <Compile Include="Tests\Legacy-CellsUITests.cs" />
-    <Compile Include="Tests\DatePickerUITests.cs" />
-    <Compile Include="Tests\BoxViewUITests.cs" />
-    <Compile Include="Tests\EditorUITests.cs" />
-    <Compile Include="Tests\FrameUITests.cs" />
-    <Compile Include="Tests\ImageUITests.cs" />
-    <Compile Include="Tests\WebViewUITests.cs" />
-    <Compile Include="Tests\TimePickerUITests.cs" />
-    <Compile Include="Tests\SwitchUITests.cs" />
-    <Compile Include="Tests\StepperUITests.cs" />
-    <Compile Include="Tests\SliderUITests.cs" />
-    <Compile Include="Tests\SearchBarUITests.cs" />
-    <Compile Include="Tests\ProgressBarUITests.cs" />
-    <Compile Include="Tests\PickerUITests.cs" />
-    <Compile Include="Tests\LabelUITests.cs" />
-    <Compile Include="Tests\EntryUITests.cs" />
-    <Compile Include="Tests\ButtonUITests.cs" />
-    <Compile Include="Tests\ViewUITests.cs" />
-    <Compile Include="Utilities\AppExtensions.cs" />
-    <Compile Include="Utilities\Drag.cs" />
-    <Compile Include="Utilities\Gestures.cs" />
-    <Compile Include="Utilities\Logger.cs" />
-    <Compile Include="Utilities\NumericExtensions.cs" />
-    <Compile Include="Utilities\ParsingUtils.cs" />
-    <Compile Include="Utilities\UITestCustomExceptions.cs" />
-    <Compile Include="Utilities\ViewInspector.cs" />
-    <Compile Include="Tests\DisplayAlertUITests.cs" />
-    <Compile Include="Tests\ActionSheetUITests.cs" />
-    <Compile Include="Tests\ToolbarItemTests.cs" />
-    <Compile Include="Tests\RootGalleryUITests.cs" />
-    <Compile Include="Tests\ScrollViewUITests.cs" />
-    <Compile Include="Tests\Legacy-UnevenListTests.cs" />
-    <Compile Include="Tests\AppearingUITests.cs" />
-    <Compile Include="Tests\AutomationIDUITests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
@@ -137,6 +91,7 @@
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
   </ItemGroup>
   <Import Project="..\Xamarin.Forms.Controls.Issues\Xamarin.Forms.Controls.Issues.Shared\Xamarin.Forms.Controls.Issues.Shared.projitems" Label="Shared" />
+  <Import Project="..\Xamarin.Forms.Core.UITests.Shared\Xamarin.Forms.Core.UITests.projitems" Label="Shared" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
